+++ tests/capture/autograph/test_autograph.py
@@ -1,0 +1,388 @@
+# Copyright 2023 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""PyTests for the AutoGraph source-to-source transformation feature."""
+
+import numpy as np
+import pytest
+
+import pennylane as qml
+from pennylane import grad, jacobian, measure
+
+pytestmark = pytest.mark.jax
+
+jax = pytest.importorskip("jax")
+
+# must be below jax importorskip
+# pylint: disable=wrong-import-position
+from pennylane.capture.autograph.transformer import TRANSFORMER, autograph_source, run_autograph
+from pennylane.capture.autograph.utils import AutoGraphError
+
+check_cache = TRANSFORMER.has_cache
+
+# pylint: disable=too-few-public-methods, unnecessary-lambda-assignment
+
+
+@pytest.fixture(autouse=True)
+def enable_disable_plxpr():
+    qml.capture.enable()
+    yield
+    qml.capture.disable()
+
+
+class TestIntegration:
+    """Test that the autograph transformations trigger correctly in different settings."""
+
+    def test_unsupported_object(self):
+        """Check the error produced when attempting to convert an unsupported object (neither of
+        QNode, function, method or callable)."""
+
+        class FN:
+            """Test object."""
+
+            __name__ = "unknown"
+
+        fn = FN()
+
+        with pytest.raises(AutoGraphError, match="Unsupported object for transformation"):
+            run_autograph(fn)
+
+    def test_callable_object(self):
+        """Test qjit applied to a callable object."""
+
+        class FN:
+            """Test object."""
+
+            __name__ = "unknown"
+
+            def __call__(self, x):
+                return x**2
+
+        fn = FN()
+
+        assert run_autograph(fn)(3) == 9
+
+    def test_lambda(self):
+        """Test autograph on a lambda function."""
+
+        fn = lambda x: x**2
+        ag_fn = run_autograph(fn)
+
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(fn)
+        assert fn(4) == 16
+
+    def test_classical_function(self):
+        """Test autograph on a purely classical function."""
+
+        def fn(x):
+            return x**2
+
+        ag_fn = run_autograph(fn)
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(fn)
+        assert fn(4) == 16
+
+    def test_nested_function(self):
+        """Test autograph on nested classical functions."""
+
+        def inner(x):
+            return x**2
+
+        def fn(x: int):
+            return inner(x)
+
+        ag_fn = run_autograph(fn)
+        assert ag_fn(4) == 16
+
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(fn)
+        assert check_cache(inner)
+
+    def test_qnode(self):
+        """Test autograph on a QNode."""
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def circ(x: float):
+            qml.RY(x, wires=0)
+            return qml.expval(qml.PauliZ(0))
+
+        ag_fn = run_autograph(circ)
+        assert ag_fn(np.pi) == -1
+
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(circ.func)
+
+    def test_indirect_qnode(self):
+        """Test autograph on a QNode called from within a classical function."""
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def inner(x):
+            qml.RY(x, wires=0)
+            return qml.expval(qml.PauliZ(0))
+
+        def fn(x: float):
+            return inner(x)
+
+        ag_fn = run_autograph(fn)
+        assert ag_fn(np.pi) == -1
+
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(fn)
+        assert check_cache(inner.func)
+
+    def test_multiple_qnode(self):
+        """Test autograph on multiple QNodes called from different classical functions."""
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def inner1(x):
+            qml.RY(x, wires=0)
+            return qml.expval(qml.PauliZ(0))
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def inner2(x):
+            qml.RX(x, wires=0)
+            return qml.expval(qml.PauliZ(0))
+
+        def fn(x: float):
+            return inner1(x) + inner2(x)
+
+        ag_fn = run_autograph(fn)
+        assert ag_fn(np.pi) == -2
+
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(fn)
+        assert check_cache(inner1.func)
+        assert check_cache(inner2.func)
+
+    def test_adjoint_wrapper(self):
+        """Test conversion is happening successfully on functions wrapped with 'adjoint'."""
+
+        def inner(x):
+            qml.RY(x, wires=0)
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def circ(x: float):
+            inner(x * 2)
+            qml.adjoint(inner)(x)
+            return qml.probs()
+
+        ag_fn = run_autograph(circ)
+        phi = np.pi / 2
+        assert np.allclose(ag_fn(phi), [np.cos(phi / 2) ** 2, np.sin(phi / 2) ** 2])
+
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(circ.func)
+        assert check_cache(inner)
+
+    def test_ctrl_wrapper(self):
+        """Test conversion is happening successfully on functions wrapped with 'ctrl'."""
+
+        def inner(x):
+            qml.RY(x, wires=0)
+
+        @qml.qnode(qml.device("default.qubit", wires=2))
+        def circ(x: float):
+            qml.PauliX(1)
+            qml.ctrl(inner, control=1)(x)
+            return qml.probs()
+
+        ag_fn = run_autograph(circ)
+        assert np.allclose(ag_fn(np.pi), [0.0, 0.0, 0.0, 1.0])
+
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(circ.func)
+        assert check_cache(inner)
+
+    def test_grad_wrapper(self):
+        """Test conversion is happening successfully on functions wrapped with 'grad'."""
+
+        def inner(x):
+            return 2 * x
+
+        def fn(x: float):
+            return grad(inner)(x)
+
+        ag_fn = run_autograph(fn)
+        assert ag_fn(3.0) == 2.0
+
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(fn)
+        assert check_cache(inner)
+
+    def test_jacobian_wrapper(self):
+        """Test conversion is happening successfully on functions wrapped with 'jacobian'."""
+
+        def inner(x):
+            return 2 * x, x**2
+
+        def fn(x: float):
+            return jacobian(inner)(x)
+
+        ag_fn = run_autograph(fn)
+        assert ag_fn(3.0) == tuple([jax.numpy.array(2.0), jax.numpy.array(6.0)])
+
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(fn)
+        assert check_cache(inner)
+
+    @pytest.mark.xfail(reason="decorated transforms are not applied yet with capture enabled")
+    def test_tape_transform(self):
+        """Test if tape transform is applied when autograph is on."""
+
+        dev = dev = qml.device("default.qubit", wires=1)
+
+        @qml.transform
+        def my_quantum_transform(tape):
+            raise NotImplementedError
+
+        def fn(x):
+            @my_quantum_transform
+            @qml.qnode(dev)
+            def circuit(x):
+                qml.RY(x, wires=0)
+                qml.RX(x, wires=0)
+                return qml.expval(qml.PauliZ(0))
+
+            return circuit(x)
+
+        ag_fn = run_autograph(fn)
+
+        with pytest.raises(NotImplementedError):
+            ag_fn(0.5)
+
+    def test_mcm_one_shot(self):
+        """Test if mcm one-shot miss transforms."""
+        dev = qml.device("default.qubit", wires=5, shots=20)
+
+        @qml.qnode(dev, mcm_method="one-shot", postselect_mode="hw-like")
+        def circ(x):
+            qml.RX(x, wires=0)
+            measure(0, postselect=1)
+            return qml.sample(wires=0)
+
+        ag_fn = run_autograph(circ)
+        # If transforms are missed, the output will be all ones.
+        assert not np.all(ag_fn(0.9) == 1)
+
+        assert hasattr(ag_fn, "ag_unconverted")
+        assert check_cache(circ.func)
+
+
+class TestCodePrinting:
+    """Test that the transformed source code can be printed in different settings."""
+
+    def test_unconverted(self):
+        """Test printing on an unconverted function."""
+
+        def fn(x):
+            return x**2
+
+        with pytest.raises(AutoGraphError, match="function was not converted by AutoGraph"):
+            autograph_source(fn)
+
+    def test_lambda(self):
+        """Test printing on a lambda function."""
+
+        fn = lambda x: x**2
+        fn = run_autograph(fn)
+
+        assert "ag__lam" in autograph_source(fn)
+
+    def test_classical_function(self):
+        """Test printing on a purely classical function."""
+
+        def fn(x):
+            return x**2
+
+        fn = run_autograph(fn)
+        assert "def ag__fn(x" in autograph_source(fn)
+
+    def test_nested_function(self):
+        """Test printing on nested classical functions."""
+
+        def inner(x):
+            return x**2
+
+        def fn(x: int):
+            return inner(x)
+
+        fn = run_autograph(fn)
+
+        # if we don't call the function, the inner function isn't found in the TRANSFORMER cache,
+        # and we can't get the source for the inner function
+        _ = fn(2)
+
+        assert "def ag__fn(x" in autograph_source(fn)
+        assert "def ag__inner(x" in autograph_source(inner)
+
+    def test_qnode(self):
+        """Test printing on a QNode."""
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def fn(x: float):
+            qml.RY(x, wires=0)
+            return qml.expval(qml.PauliZ(0))
+
+        fn = run_autograph(fn)
+
+        assert autograph_source(fn)
+
+    def test_indirect_qnode(self):
+        """Test printing on a QNode called from within a classical function."""
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def inner(x):
+            qml.RY(x, wires=0)
+            return qml.expval(qml.PauliZ(0))
+
+        def fn(x: float):
+            return inner(x)
+
+        fn = run_autograph(fn)
+        # if we don't call the function, the inner function isn't found in the TRANSFORMER cache,
+        # and we can't get the source for the inner function
+        _ = fn(2)
+
+        assert "def ag__fn(x" in autograph_source(fn)
+        assert "def ag__inner(x" in autograph_source(inner)
+
+    def test_multiple_qnode(self):
+        """Test printing on multiple QNodes called from different classical functions."""
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def inner1(x):
+            qml.RY(x, wires=0)
+            return qml.expval(qml.PauliZ(0))
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def inner2(x):
+            qml.RX(x, wires=0)
+            return qml.expval(qml.PauliZ(0))
+
+        def fn(x: float):
+            return inner1(x) + inner2(x)
+
+        fn = run_autograph(fn)
+        # if we don't call the function, the inner function isn't found in the TRANSFORMER cache,
+        # and we can't get the source for the inner function
+        _ = fn(2)
+
+        assert "def ag__fn(x" in autograph_source(fn)
+        assert "def ag__inner1(x" in autograph_source(inner1)
+        assert "def ag__inner2(x" in autograph_source(inner2)
+
+
+if __name__ == "__main__":
+    pytest.main(["-x", __file__])
