+++ tests/capture/autograph/test_conditionals.py
@@ -1,0 +1,282 @@
+# Copyright 2023 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""PyTests for the AutoGraph source-to-source transformation feature."""
+
+# pylint: disable=wrong-import-order, wrong-import-position, ungrouped-imports
+
+import pytest
+
+import pennylane as qml
+from pennylane import cond, measure
+
+pytestmark = pytest.mark.jax
+
+jax = pytest.importorskip("jax")
+
+# must be below jax importorskip
+from jax.core import eval_jaxpr
+
+from pennylane.capture.autograph.transformer import TRANSFORMER, run_autograph
+from pennylane.capture.autograph.utils import AutoGraphError
+
+check_cache = TRANSFORMER.has_cache
+
+
+@pytest.fixture(autouse=True)
+def enable_disable_plxpr():
+    qml.capture.enable()
+    yield
+    qml.capture.disable()
+
+
+class TestConditionals:
+    """Test that the autograph transformations produce correct results on conditionals.
+    These tests are adapted from the test_conditionals.TestCond class of tests."""
+
+    def test_simple_cond(self):
+        """Test basic function with conditional."""
+
+        def circuit(n):
+            if n > 4:
+                res = n**2
+            else:
+                res = n
+
+            return res
+
+        # can't convert to jaxpr without autorgraph
+        with pytest.raises(
+            jax.errors.TracerBoolConversionError,
+            match="Attempted boolean conversion of traced array",
+        ):
+            jax.make_jaxpr(circuit)(1)
+
+        # with autograph we can convert to jaxpr
+        circuit = run_autograph(circuit)
+        jaxpr = jax.make_jaxpr(circuit)(0)
+        assert "cond" in str(jaxpr)
+
+        def res(x):
+            return eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, x)
+
+        # evaluating the jaxpr gives expected results
+        assert res(0) == [0]
+        assert res(1) == [1]
+        assert res(2) == [2]
+        assert res(3) == [3]
+        assert res(4) == [4]
+        assert res(5) == [25]
+        assert res(6) == [36]
+
+    def test_cond_one_else_if(self):
+        """Test a cond with one else_if branch"""
+
+        def circuit(x):
+            if x > 2.7:
+                res = x * 4
+            elif x > 1.4:
+                res = x * 2
+            else:
+                res = x
+
+            return res
+
+        ag_circuit = run_autograph(circuit)
+        jaxpr = jax.make_jaxpr(ag_circuit)(0)
+        assert "cond" in str(jaxpr)
+
+        def res(x):
+            return eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, x)[0]
+
+        assert res(4) == 16
+        assert res(2) == 4
+        assert res(1) == 1
+
+    def test_cond_many_else_if(self):
+        """Test a cond with multiple else_if branches"""
+
+        def circuit(x):
+            if x > 4.8:
+                res = x * 8
+            elif x > 2.7:
+                res = x * 4
+            elif x > 1.4:
+                res = x * 2
+            else:
+                res = x
+
+            return res
+
+        ag_circuit = run_autograph(circuit)
+        jaxpr = jax.make_jaxpr(ag_circuit)(0)
+        assert "cond" in str(jaxpr)
+
+        def res(x):
+            return eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, x)[0]
+
+        assert res(5) == 40
+        assert res(3) == 12
+        assert res(2) == 4
+        assert res(-3) == -3
+
+    def test_qubit_manipulation_cond(self):
+        """Test conditional with quantum operation."""
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def circuit(x):
+            if x > 4:
+                qml.PauliX(wires=0)
+
+            m = measure(wires=0)
+
+            return qml.expval(m)
+
+        ag_circuit = run_autograph(circuit)
+        jaxpr = jax.make_jaxpr(ag_circuit)(0)
+        assert "cond" in str(jaxpr)
+
+        def res(x):
+            return eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, x)[0]
+
+        # pylint: disable=singleton-comparison
+        assert res(3) == 0
+        assert res(6) == 1
+
+    def test_branch_return_mismatch(self):
+        """Test that an exception is raised when the true branch returns a value without an else
+        branch.
+        """
+        # pylint: disable=using-constant-test
+
+        def circuit():
+            if True:
+                res = measure(wires=0)
+
+            return qml.expval(res)
+
+        with pytest.raises(
+            AutoGraphError, match="Some branches did not define a value for variable 'res'"
+        ):
+            qml.capture.autograph.run_autograph(circuit)()
+
+    def test_branch_multi_return_mismatch(self):
+        """Test that an exception is raised when the return types of all branches do not match."""
+        # pylint: disable=using-constant-test
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def circuit():
+            if True:
+                res = measure(wires=0)
+            elif False:
+                res = 0.0
+            else:
+                res = measure(wires=0)
+
+            return res
+
+        with pytest.raises(
+            TypeError, match="Conditional requires consistent return types across all branches"
+        ):
+            run_autograph(circuit)
+
+    def test_multiple_return(self):
+        """Test return statements from different branches of an if/else statement
+        with autograph."""
+
+        # pylint: disable=no-else-return
+
+        def f(x: int):
+            if x > 0:
+                return 25  # converted to cond_fn
+            else:
+                return 60  # converted to else_fn
+
+        ag_circuit = run_autograph(f)
+        jaxpr = jax.make_jaxpr(ag_circuit)(0)
+        assert "cond" in str(jaxpr)
+
+        def res(x):
+            return eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, x)[0]
+
+        assert res(1) == 25
+        assert res(0) == 60
+
+    def test_multiple_return_early(self):
+        """Test that returning early is possible, and that the final return outside
+        the conditional works as expected."""
+
+        def f(x: float):
+
+            if x:
+                return x  # converted to cond_fn with no else_fn
+
+            x = x + 2
+            return x
+
+        ag_circuit = run_autograph(f)
+        jaxpr = jax.make_jaxpr(ag_circuit)(0)
+
+        def res(x):
+            return eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, x)[0]
+
+        # returning early is an option, and code after the return is not executed
+        assert res(1) == 1
+
+        # if an early return isn't hit, the code between the early return and the
+        # final return is executed
+        assert res(0) == 2
+
+    def test_multiple_return_mismatched_type(self):
+        """Test that different observables cannot be used in the return in different branches."""
+
+        # ToDo: I don't see this contraint - should I?
+
+        @qml.qnode(qml.device("default.qubit", wires=1))
+        def f(switch: bool):
+            if switch:
+                return qml.expval(qml.PauliY(0))
+
+            return qml.expval(qml.PauliZ(0))
+
+        with pytest.raises(TypeError, match="requires a consistent return structure"):
+            ag_circuit = run_autograph(f)
+            jaxpr = jax.make_jaxpr(ag_circuit)(0)
+
+            def res(x):
+                return eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, x)[0]
+
+            res(1)
+
+    def test_cond(self):
+        """Test if Autograph works when applied directly to a decorated function with cond"""
+
+        n = 6
+
+        @cond(n > 4)
+        def cond_fn():
+            return n**2
+
+        @cond_fn.otherwise
+        def else_fn():
+            return n
+
+        ag_fn = run_autograph(cond_fn)
+        jaxpr = jax.make_jaxpr(ag_fn)()
+
+        assert eval_jaxpr(jaxpr.jaxpr, jaxpr.consts)[0] == 36
+
+
+if __name__ == "__main__":
+    pytest.main(["-x", __file__])
